Rename seviri readers to include 'l1b' in the name
