+++ satpy/readers/seviri_base.py
