+++ satpy/readers/seviri_l1b_native.py
@@ -45,11 +45,11 @@ from pyresample import geometry
 
 from satpy.readers.file_handlers import BaseFileHandler
 from satpy.readers.eum_base import recarray2dict
-from satpy.readers.msg_base import (SEVIRICalibrationHandler,
-                                    CHANNEL_NAMES, CALIB, SATNUM,
-                                    dec10216)
-from satpy.readers.seviri_native_hdr import (GSDTRecords, native_header,
-                                             native_trailer)
+from satpy.readers.seviri_base import (SEVIRICalibrationHandler,
+                                       CHANNEL_NAMES, CALIB, SATNUM,
+                                       dec10216)
+from satpy.readers.seviri_l1b_native_hdr import (GSDTRecords, native_header,
+                                                 native_trailer)
 
 
 logger = logging.getLogger('native_msg')
